+++ docs/changelog.rst
@@ -9,7 +9,12 @@ Changelog
 Added
 ~~~~~
 
-- Display progress during the CLI run `#125`_
+- Display progress during the CLI run. `#125`_
+
+Fixed
+~~~~~
+
+- Test server generated wrong schema when ``endpoints`` option is passed via CLI. `#173`_
 
 `0.11.0`_ - 2019-10-22
 ----------------------
@@ -264,6 +269,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#173: https://github.com/kiwicom/schemathesis/issues/173
 .. _#167: https://github.com/kiwicom/schemathesis/issues/167
 .. _#153: https://github.com/kiwicom/schemathesis/issues/153
 .. _#149: https://github.com/kiwicom/schemathesis/issues/149